fix pdf rendering for hidpi with fractional device pixel ratio

+++ src/pdfviewer/PDFDocument.h
@@ -80,7 +80,7 @@ protected:
 private:
 	int oldshape;
 	int page;
-	int overScale;
+    qreal   overScale;
 	qreal	scaleFactor;
 	qreal	parentDpi;
 	QPixmap	image;
